attempt to cache api
The bikeshare api sets the no-cache header, so in order to cache
at the level of the HTTP request, you have to intercept the response
and re-write the header.  (at least that's my current understanding.)
This is something that Jake Wharton calls "an extreme measure", that's
more or less a direct quote.

I couldn't get this working to my satisfaction (how to log a cache hit or
cache miss wasn't immediately obvious and I'm still seeing GET requests
even when I think the cache should be fresh.

I commented out the code, but I'm leaving it in there for now for future
reference.